requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.4.2
chromadb>=0.4.18
promptlayer>=0.2.5
mlflow>=2.8.1
pypdf>=3.17.1
python-multipart>=0.0.6